Move `getEnvironmentVariable` and Terminal to `Environment.swift`.
