+++ Source/carthage/CopyFrameworks.swift
@@ -94,13 +94,3 @@ private func inputFiles() -> ColdSignal<String> {
 			return ColdSignal.fromValues(variables).concat(identity)
 		}
 }
-
-internal func getEnvironmentVariable(variable: String) -> Result<String> {
-	let environment = NSProcessInfo.processInfo().environment
-
-	if let value = environment[variable] as String? {
-		return success(value)
-	} else {
-		return failure(CarthageError.MissingEnvironmentVariable(variable: variable).error)
-	}
-}
